+++ Samples/macOS-Swift/macOS-Swift/AppDelegate.swift
@@ -10,6 +10,8 @@ class AppDelegate: NSObject, NSApplicationDelegate {
             options.dsn = "https://a92d50327ac74b8b9aa4ea80eccfb267@o447951.ingest.sentry.io/5428557"
             options.debug = true
             options.sessionTrackingIntervalMillis = 5_000
+            // Sampling 100% - In Production you probably want to adjust this
+            options.tracesSampleRate = 1.0
         }
         
         SentrySDK.configureScope { scope in