+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased 
 
 - fix: Race on session task (#1233)
+- fix: ViewController swizzling before iOS 13 (#1231)
 - fix: AppStartMeasurement didFinishLaunching is nil (#1228)
 
 ## 7.2.0-beta.5